[LED-69] Extracted TransactionModel class, renamed the pending/cleared column.
