+++ src/main/java/ledger/user_interface/ui_controllers/MainPageController.java
@@ -26,6 +26,7 @@ import ledger.database.entity.Tag;
 import ledger.database.entity.Transaction;
 import ledger.database.entity.Type;
 import ledger.exception.StorageException;
+import ledger.user_interface.ui_models.TransactionModel;
 
 import java.net.URL;
 import java.text.DateFormat;
@@ -67,7 +68,7 @@ public class MainPageController extends GridPane implements Initializable {
     @FXML
     private TableColumn categoryColumn;
     @FXML
-    private TableColumn closedColumn;
+    private TableColumn clearedColumn;
 
     // Transaction table edit event handlers
     private EventHandler<CellEditEvent<TransactionModel, String>> amountEditHandler = new EventHandler<CellEditEvent<TransactionModel, String>>() {
@@ -346,7 +347,7 @@ public class MainPageController extends GridPane implements Initializable {
         this.payeeColumn.setCellValueFactory(new PropertyValueFactory<TransactionModel, String>("payeeName"));
         this.typeColumn.setCellValueFactory(new PropertyValueFactory<TransactionModel, String>("typeName"));
         this.categoryColumn.setCellValueFactory(new PropertyValueFactory<TransactionModel, String>("tagNames"));
-        this.closedColumn.setCellValueFactory(new PropertyValueFactory<TransactionModel, String>("pending"));
+        this.clearedColumn.setCellValueFactory(new PropertyValueFactory<TransactionModel, String>("pending"));
 
         this.amountColumn.setCellFactory(TextFieldTableCell.forTableColumn());
         this.amountColumn.setOnEditCommit(this.amountEditHandler);
@@ -363,8 +364,8 @@ public class MainPageController extends GridPane implements Initializable {
         this.categoryColumn.setCellFactory(TextFieldTableCell.forTableColumn());
         this.categoryColumn.setOnEditCommit(this.categoryEditHandler);
 
-        this.closedColumn.setCellFactory(TextFieldTableCell.forTableColumn());
-        this.closedColumn.setOnEditCommit(this.closedEditHandler);
+        this.clearedColumn.setCellFactory(TextFieldTableCell.forTableColumn());
+        this.clearedColumn.setOnEditCommit(this.closedEditHandler);
 
         // Add ability to delete transactions form tableView
         this.transactionTableView.setOnKeyPressed(new EventHandler<KeyEvent>() {
@@ -414,121 +415,6 @@ public class MainPageController extends GridPane implements Initializable {
             }
         } catch (StorageException e) {
             System.out.println("Error deleting transaction: " + e.getMessage());
-        }
-    }
-
-    public class TransactionModel {
-        private int id;
-        private String amount;
-        private String date;
-        private String payeeName;
-        private String typeName;
-        private String tagNames;
-        private String pending;
-        private Transaction transaction;
-
-        public TransactionModel(Transaction transaction) {
-            this.id = transaction.getId();
-            String amountInCents = String.valueOf(transaction.getAmount());
-            String dollars = amountInCents.substring(0, amountInCents.length() - 2);
-            String cents = amountInCents.substring(amountInCents.length() - 2, amountInCents.length());
-            this.amount = "$" + dollars + "." + cents;
-            if (transaction.getDate() != null) {
-                this.date = transaction.getDate().toString();
-            } else {
-                this.date = "";
-            }
-            if (transaction.getPayee() != null) {
-                this.payeeName = transaction.getPayee().getName();
-            } else {
-                this.payeeName = "";
-            }
-            if (transaction.getType() != null) {
-                this.typeName = transaction.getType().getName();
-            } else {
-                this.typeName = "";
-            }
-            this.tagNames = "";
-            if (transaction.getTagList() != null) {
-                List<Tag> tags = transaction.getTagList();
-                for (int i = 0; i < tags.size(); i++) {
-                    this.tagNames += tags.get(i).getName();
-                    if (i != tags.size() - 1) {
-                        this.tagNames += ", ";
-                    }
-                }
-            }
-            if (transaction.isPending()) {
-                this.pending = "Pending";
-            } else {
-                this.pending = "Cleared";
-            }
-
-            this.transaction = transaction;
-        }
-
-        public int getId() {
-            return id;
-        }
-
-        public void setId(int id) {
-            this.id = id;
-        }
-
-        public String getAmount() {
-            return amount;
-        }
-
-        public void setAmount(String amount) {
-            this.amount = amount;
-        }
-
-        public String getDate() {
-            return date;
-        }
-
-        public void setDate(String date) {
-            this.date = date;
-        }
-
-        public String getPayeeName() {
-            return payeeName;
-        }
-
-        public void setPayeeName(String payeeName) {
-            this.payeeName = payeeName;
-        }
-
-        public String getTypeName() {
-            return typeName;
-        }
-
-        public void setTypeName(String typeName) {
-            this.typeName = typeName;
-        }
-
-        public String getTagNames() {
-            return tagNames;
-        }
-
-        public void setTagNames(String tagNames) {
-            this.tagNames = tagNames;
-        }
-
-        public String getPending() {
-            return pending;
-        }
-
-        public void setPending(String pending) {
-            this.pending = pending;
-        }
-
-        public Transaction getTransaction() {
-            return transaction;
-        }
-
-        public void setTransaction(Transaction transaction) {
-            this.transaction = transaction;
         }
     }
 }
